Modernize public API: remove C type leakage, consistent error handling, clean up demo state
- Replace UInt32/Int32 with Int across public API (Track, Equalizer, Player, Media, Metadata, ThumbnailRequest)
- Convert Bool-returning methods to throws(VLCError) (Player, MediaList, MediaListPlayer, Media, Equalizer)
- Convert setter methods to properties (Player.equalizer, MediaListPlayer.mediaPlayer/mediaList/playbackMode, Marquee.text, Logo.file)
- Rename addSlave → addExternalTrack, add play(Media) and play(url:) convenience methods
- Add Duration.formatted, VLCError.operationFailed, Equatable on Metadata/MediaStatistics
- Label PlayerEvent.recordingChanged tuple params, name RendererDiscoverer flag constants
- Remove redundant @State in demo (volume, equalizer, eqEnabled, liveStats) — use player properties directly

+++ Sources/SwiftVLC/Media/ThumbnailRequest.swift
@@ -19,8 +19,8 @@ public extension Media {
     /// - Returns: The raw image data (PNG format).
     func thumbnail(
         at time: Duration,
-        width: UInt32 = 320,
-        height: UInt32 = 0,
+        width: Int = 320,
+        height: Int = 0,
         crop: Bool = false,
         timeout: Duration = .seconds(10),
         instance: VLCInstance = .shared
@@ -45,8 +45,8 @@ public extension Media {
                 media,
                 time.milliseconds,
                 libvlc_media_thumbnail_seek_fast,
-                width,
-                height,
+                UInt32(width),
+                UInt32(height),
                 crop,
                 libvlc_picture_Png,
                 timeout.milliseconds
